Renamed types to interfaces.
