+++ Frontend/src/interfaces/carousel.interface.tsx
