+++ Frontend/src/interfaces/customer.interface.tsx
@@ -1,4 +1,4 @@
-export interface Customer {
+export interface ICustomer {
   firstName: string
   email: string
   phoneNumber: string
