+++ Frontend/src/interfaces/defaultPage.interface.tsx
@@ -1,0 +1,14 @@
+import { IHeader } from "./header.interface"
+import { ICarousel } from "./carousel.interface"
+import { IFooter } from "./footer.interface"
+
+export interface IDefaultPage {
+  title: string
+  slug: string
+  header: IHeader
+  type: "none" | "category"
+  category: string
+  carousel: ICarousel
+  description: string
+  footer: IFooter
+}
