+++ Frontend/src/interfaces/footer.interface.tsx
