+++ Frontend/src/interfaces/route.interface.tsx
@@ -1,0 +1,5 @@
+export interface IRoute {
+  title: string
+  slug: string
+  type: "category" | "none"
+}
