--- Frontend/src/types/route.type.tsx
@@ -1,5 +1,0 @@
-export interface Route {
-  title: string
-  slug: string
-  type: "category" | "tag" | "none"
-}
